Add schema.js and session-management.service
